relocating api manager and rotting
